fixed average price calculation

+++ intellicator.h
@@ -20,7 +20,7 @@ public slots:
     void startTask();
     void startSimulation();
     void updateProgress(int value);
-    void finishedTask(int value);
+    void finishedTask(double value);
     void createChart(const std::vector<Point>& data);
 
 private:
